setup pixijs, got some bunnies moving aroudn and sending back data, friction based movement needs to be synced with other clients

+++ entities/payload.go
@@ -1,7 +1,7 @@
 package entities
 
 type Payload struct {
-	X    int    `json:"x"`
-	Y    int    `json:"y"`
-	Uuid string `json:"uuid"`
+	X    float64 `json:"x"`
+	Y    float64 `json:"y"`
+	Uuid string  `json:"uuid"`
 }
